Add header support for requests/responses
Note: we have a bit of a hack for raw headers, where we use a _raw_ key.
This is because I want the internal structure for Headers to be
map[string]string since that is the most common use, but want to support
sending the headers as is for the raw encoding over TChannel.