thrift: Improve handling of enums
When an enum value is received over the wire, it is formatted using the
name. If a name is not found, it is formatted as EnumType(%d)

Enums can be specified as:
- int32
- Case-insentivie name
- EnumType(%d)

The last format is so that formatted unknown enums on the result path
can be used as input.