Release latest yab changes as 0.2.0 (#43)
* Increment version for dev branch

* Add yaml to glide.yaml and glide.lock

* Use YAML instead of JSON when parsing

YAML allows for more types than JSON (e.g. non-string keys) but
is also backwards compatible so any JSON string the user passes
can be parsed.

* Add integration tests for yaml/json -> thrift -> yaml

* YAML everywhere (and update options documentation)

* Refactor encoding code into separate package

* Add aliases -2 and -3 for headers/body to match tcurl

* Add a hidden option to display a man page in groff (#35)

* Fix bad merge

* Use in-memory filesystem for parsing test thrift files

* thrift: Improve handling of enums

When an enum value is received over the wire, it is formatted using the
name. If a name is not found, it is formatted as EnumType(%d)

Enums can be specified as:
- int32
- Case-insentivie name
- EnumType(%d)

The last format is so that formatted unknown enums on the result path
can be used as input.

* thrift/to_wire: Tests (#39)

This changes the ValueFromWire success tests to test verify the ToWire
direction too.

* thrift: Resolve typedefs before wire conversion (#38)

* Update thriftrw, always set missing fields to defaults (#40)

* Use NonStrict parsing in yab (#41)

* Use compile.RootTypeSpec in latest thriftrw-go (#42)

* Update version to 0.2.0