benchmark: Round robin peer selection to balance connections
Previously, we relied on randomly picking a peer for each connection
which could lead to an imbalance. E.g., if there are 3 peers and 3
connections, there was no guarantee that all 3 peers would receive
a single connection.

Instead of picking randomly, round-robin the peers that can be used
for each connection (starting at a random offset) for better load
balancing.