Fix benchmark taking longer than duration for low RPS
When the RPS is a low number (like 1), each connection goroutine will
call Take which ends up causing a sleep for a second. Since the limiter
uses a lock, we end up serializing the Sleeps, so the benchmark runs for
connections * sleep amount.

Instead of sleeping arbitrarily, pass in a cancel channel to unblock the
sleep.