Override default peer options based on args
If defaults sets the hosts file, but args sets a single peer, the
final options should only contain the single peer. If both are set,
yab returns an error.

This is required if we want to set a default peer list in a global
config that is only used when the user doesn't override the peer in
any way.