try to check the system xdg dir for a config file
This tests the user's config dir for a yab config, and if that fails yab
will fallback to checking the system directories. On a Linux machine
this means that the config files will be checked in this order (with the
first found taking precedence):

    - ~/.config/yab/defaults.ini
    - /etc/xdg/yab/defaults.ini

The motivation behind this change is to allow system administrators to
install a system-wide default configuration file.