benchmarking: Allow unlimited requests (capped by max duration) (#111)
Currently we always require a MaxRequests. Instead change the default to
have no MaxRequests, and make only MaxRequests or MaxDuration required
for a benchmark.

If RPS and MaxDuration are provided, an implicit MaxRequests is
calculated. Otherwise, the test runs until the max duration is hit.

This change also disallows non-negative max requests/durations.