Add support for template arguments
Template arguments are of the format ${argname:fallback}. Arguments can
be specified on the commandline using `--arg` or `-A`. Arguments are
parsed as YAML values, and allow more complex types (e.g., lists) to be
specified as arguments.

We only use template arguments for request bodies currently.

Argument parsing is based on @abhinav's work for YARPC configuration.