Fix parsing of strings from values that look like ints/bools (#174)
Since YAML tries to guess the type of a value, a string argument with a
value that looks like an int or a bool will not be a string type. We
should coerce these values into a string to users are not forced to
explicitly quote strings in their YAML file.