Fix timeouts in templates being ignored (#210)
We parse YAML templates before flags so that users can override template
args using flags. This didn't work for timeout since we used the
`default` flag, which will always set the option in the struct if the
flag was not specified. This caused the timeout value read from YAML to
be overridden.

We can avoid go-flags from setting the value by setting the default
value on the struct. go-flags displayed this default as an integer, so
use default-mask to display it as a human readable string.