Support yab as a shebang without -y
Shebangs cannot contain flags on Linux, so our recommendation to use
    #!/usr/bin/env yab -y
doesn't work on Linux.

Instead, have yab look at the first argument, and if it's a valid file
ending in .yab, then don't require -y. This is a little bit of "magic"
but will make the user experience better (so users don't have to
remember to use a separate script called yaby or something).

Fixes #198.