protobuf: Stop reflection client for graceful stream shutdown (#247)
Currently, the reflection stream is closed abruptly when yab ends,
which looks like an error for intermediate proxies.

Add a Close method to the reflection source API and ensure we reset
the grpc client which stops the underlying stream to avoid this error.

For testing, use the grpc GracefulStop API, which blocks until all RPCs
are complete. If the test times out, then the reflection stream isn't
being shutdown correctly.