thrift: Treat null map values as ignore field/key (#261)
Currently, if a value is specified as null, it causes an error
parsing (e.g., cannot parse null as i32). This limits how useful
yab templates can be, as any specified field must always be set to
a value.

However, what we'd like is a YAML template with something like:

```
request:
  minVersion: ${minVersion:null}
  maxVersion: ${maxVersion:null}
```

We may only want to specify one or the other condition, but today this
template would cause errors and it would try to parse "null". Instead we
should ignore null, as null is not a value we can propagate on the wire.