benchmark: Add per-peer stats
When a list of backends is slow, it's hard to figure out which backend is
slow when using the current metrics (which are aggregates).

Add a per-peer-stats option that will track metrics by peer rather than
aggregated together. This makes it significantly easier to find a single
slow peer.

This adds a few statsd utilities to add prefixes, dual emit, and to
simplify testing.