Refactor transport & encoding detection
The code to detect encoding/protocol is split up into phases, since we
have to create/detect one before the other.

This has caused issues (e.g., unspecifiedHealthSeriailizer) which then
led to issues like https://github.com/yarpc/yab/issues/281

Rather than fix these as one-offs, refactor the code so we try to
determine the transport protocol & encoding together as one step before
we create the transport or the serializer.

This should make it easier to reason about which encoding/transport
we'll use when the user doesn't specify one.

Fixes #281.