encoding: Improve error message for unsupported encoding with --health (#283)
In previous versions of yab, `--health` only supported Thrift. Now that
we support Protobuf, we default to the right encoding based on
transport, see #282.

For TChannel, we assume thrift
For gRPC, we assume proto
For HTTP, we assume JSON

The latter used to presume Thrift, but now will just fail. In future, we
may want to use `/health` for HTTP --health, so rather than going back
to Thrift, give a clearer error message which tells users how to specify
the encoding explicitly.