Introduce yamlalias to simplify aliases in templates (#263)
For yab templates, we support a few aliases to simplify
usage for users. E.g., we support shardKey, shardkey, shard-key or sk.

This requires a lot of manual wiring up to work. Instead, introduce
a yamlalias package which will use reflection to generate a new struct
that contains additional fields for each alias.