Provide proper support for responses that contain protobuf.Any
By default protoreflect fails when there is an any type and it's type
url is unkonwn to in the extension registry. This change tries to use
the descriptor providers that we have registered in yab (if provided)
otherwise it will just use a simple bytes wrapper to represent the any
type. This results in a base64 encoding in yab requests.

See https://github.com/golang/protobuf/issues/747 and
https://github.com/protocolbuffers/protobuf/issues/5226 for more context
on what this task is fixing