Update jaeger-client-go, drop uber-go/atomic (#355)
The github.com/uber-go/atomic import path was deprecated in favor of
go.uber.org/atomic in [v1.5.0 of atomic][1], in 2019.

[1]: https://github.com/uber-go/atomic/blob/master/CHANGELOG.md#150---2019-10-29

jaeger-client-go was the last remaining straggler because of which we
still needed that import path and the corresponding replace directive.
(Determined with `go mod why github.com/uber/jaeger-client-go`.)

This updates the dependency and drops the `replace` directive.

Depends on #354