cactus/go-statsd-client: Upgrade to v5 (#356)
The version of the go-statsd-client we're using is woefully old.
Upgrade to the latest version (v5).
There's a [changelog here][1], but the only change that affects us is
that the new NewNoopClient constructor was deleted.
A nil `*statsd.Client` is a no-op client instead.

[1]: https://github.com/cactus/go-statsd-client/blob/master/CHANGELOG.md

The deleted ensureStatsd function existed only to ensure that the
`NewNoopClient() (Statter, error)` function's error was handled cleanly.
Since constructing a nil `*statsd.Client` is free and will not fail, the
function is no longer necessary.

Refs #355