CI: Add dependabot
Use dependabot to upgrade dependencies for yab.

We don't want to use dependabot for our libraries because it forces
consumers of those libraries to always upgrade all transitive
dependencies, even if not yet necessary for us.

However, since yab is intended to be a binary, not a library,
it's okay to do this in yab.