Output errors in JSON format
So far, the errors encountered during a YAB benchmark run would be printed as text and not json, when format is json. This PR fixes that.

I'm rearranging the error printing a bit to align it with the other info printed. Lastly, I've added a test that verifies the output when errors are encountered (something that was not tested before).