Fix broken in tests
In https://github.com/grpc/grpc-go/pull/5197, there was a major change
to server reflection, specifically:

"If a [DescriptorResolver] is not present, protoregistry.GlobalFiles
will be used. "

It so happens that in simple.pb in the testdata there *is* a service
called Foo and Bar and this is why these tests are now surprisingly
passing instead of failing with unknown method errors.

We work around this by creating method names that are unlikely to exist.